Move restrictions to domain folder
