+++ packages/app/src/domain/restrictions/lockdown-table.tsx
