+++ packages/app/src/domain/restrictions/restriction-icons.ts
