+++ packages/app/src/utils/use-escalation-color.ts
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { regionThresholds } from '~/components/choropleth/region-thresholds';
-import { EscalationLevel } from '~/components/restrictions/type';
+import { EscalationLevel } from '~/domain/restrictions/type';
 import { assert } from './assert';
 
 export function useEscalationColor(level?: EscalationLevel) {
